Creating an intermediary Metadata class to implement metadata adding optimization an separete from synch method implementation.
