+++ iped-parsers/iped-parsers-impl/src/main/java/iped/parsers/jdbc/AbstractDBParser.java
@@ -40,7 +40,7 @@ import org.xml.sax.SAXException;
 
 import iped.parsers.standard.StandardParser;
 import iped.properties.ExtraProperties;
-import iped.utils.tika.SyncMetadata;
+import iped.utils.tika.IpedMetadata;
 
 /**
  * Abstract class that handles iterating through tables within a database.
@@ -80,7 +80,7 @@ public abstract class AbstractDBParser extends AbstractParser {
         do {
             EmbeddedDocumentExtractor extractor = context.get(EmbeddedDocumentExtractor.class,
                     new ParsingEmbeddedDocumentExtractor(context));
-            SyncMetadata tableM = new SyncMetadata();
+            IpedMetadata tableM = new IpedMetadata();
             do {            	
                 try(InputStream is = trg.createHtmlReport(tableRowsPerItem, handler, context, tmp, tableM)) {
                     ++table_fragment;
@@ -149,10 +149,6 @@ public abstract class AbstractDBParser extends AbstractParser {
 
             for (String tableName : tableNames) {
                 JDBCTableReader reader = getTableReader(connection, tableName, context);
-                /*
-                long rowCount = reader.getRowCount();
-                metadata.set(Database.ROW_COUNT, rowCount);
-                */
                 
                 int row_count = parseTables(xHandler, context, reader);
                 xHandler.startElement("tr");
