+++ iped-parsers/iped-parsers-impl/src/main/java/iped/parsers/jdbc/MetadataColumnExtractor.java
@@ -3,13 +3,13 @@ package iped.parsers.jdbc;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
-import iped.utils.tika.SyncMetadata;
+import iped.utils.tika.IpedMetadata;
 
 public interface MetadataColumnExtractor {
 	String[] getColumnNames();
 
 	public boolean extractMetadata(ResultSet rs, int columnIndex) throws SQLException;
-	public void applyExtractedMetadatas(SyncMetadata metadata);
+	public void applyExtractedMetadatas(IpedMetadata metadata);
 	public String getMetadataName();
 	public void cancelLastExtraction();
 	public String getLastExtraction();
